server/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "transcriptionsuite-server"
version = "1.1.3"
description = "TranscriptionSuite unified server for transcription services"
requires-python = ">=3.13,<3.14"
license = "GPL-3.0-or-later"
authors = [
    { name = "homelab-00" }
]
keywords = ["transcription", "whisper", "speech-to-text", "diarization"]
classifiers = [
    "Programming Language :: Python :: 3.13",
    "Operating System :: POSIX :: Linux",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    # Web / API
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "python-multipart>=0.0.21",
    "pydantic>=2.12.5",
    "aiohttp>=3.13.3",
    "httpx>=0.28.1",
    "websockets>=15.0.1",

    # Core ML / Transcription
    "torch>=2.8.0",          # Required by PyAnnote 4.0.3
    "torchaudio>=2.8.0",     # Required by PyAnnote 4.0.3
    "pyannote.audio>=4.0.3", # Diarization (still needed for NeMo backends)

    # Audio Processing
    "numpy>=2.4.0",
    "soundfile>=0.13.1",
    "webrtcvad>=2.0.10",
    "silero-vad>=6.2.0",
    "ffmpeg-python>=0.2.0",
    "scipy>=1.17.0",

    # Database
    "aiosqlite>=0.22.1",
    "sqlalchemy>=2.0.45",
    "alembic>=1.17.1",

    # Utilities
    "setuptools<81",          # webrtcvad imports pkg_resources, removed in setuptools>=81
    "pyyaml>=6.0.3",
    "psutil>=7.2.0",
    "filelock>=3.20.1",
    "aiofiles>=24.1.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
whisper = [
    "faster-whisper>=1.2.1",
    "ctranslate2>=4.6.2",
    "whisperx>=3.1.0",       # WhisperX: faster-whisper + wav2vec2 alignment + diarization
]
nemo = [
    "nemo_toolkit[asr]>=2.2.0",
]
vibevoice_asr = [
    "vibevoice @ git+https://github.com/microsoft/VibeVoice.git@1807b858d4f7dffdd286249a01616c243e488c9e",
]
[project.scripts]
transcription-server = "server.api.main:main"

[tool.hatch.build.targets.wheel]
packages = ["server"]

[tool.hatch.metadata]
allow-direct-references = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

# The section below is needed to specifically get CUDA 12.9 (and not the default 12.8)
[[tool.uv.index]]
name = "pytorch-cu129"
url = "https://download.pytorch.org/whl/cu129"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu129" }]
torchaudio = [{ index = "pytorch-cu129" }]
